controller test ready

+++ src/main/java/org/yourcompany/yourproject/rest/UserController.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.yourcompany.yourproject.dao.UserRepository;
+import org.yourcompany.yourproject.exception.UserNotFoundException;
 import org.yourcompany.yourproject.model.User;
 
 import jakarta.validation.Valid;
@@ -29,36 +31,40 @@ public class UserController {
     }
 
     @GetMapping("")
-    List<User> findAll() {
-        return userRepository.findAll();
+    ResponseEntity<List<User>> findAll() {
+        return new ResponseEntity<>(userRepository.findAll(), HttpStatus.OK);
     }
  
     @GetMapping("/{id}")
-    User findById(@PathVariable Long id) {
+    ResponseEntity<User> findById(@PathVariable Long id) {
         Optional<User> user = userRepository.findById(id);
-
         if (user.isPresent()) {
-            return user.get();
-        } else {
-            throw new RuntimeException("User not found");
+            return new ResponseEntity<>(user.get(), HttpStatus.OK);
+        }else{
+            throw new UserNotFoundException("User not found");
         }
     }
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("")
-    void create(@Valid @RequestBody User user) {
-        userRepository.save(user);
+    ResponseEntity<User> create(@Valid @RequestBody User user) {
+        User savedUser = userRepository.save(user);
+        return new ResponseEntity<>(savedUser, HttpStatus.CREATED);
     }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PutMapping("/{id}")
-    void update(@Valid @RequestBody User user, @PathVariable Long id) throws Exception{
-        userRepository.save(user);
+    ResponseEntity<User> update(@Valid @RequestBody User user, @PathVariable Long id) throws Exception{
+        return new ResponseEntity<>(userRepository.save(user), HttpStatus.CREATED);
     }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
     void delete(@PathVariable Long id) {
-        userRepository.deleteById(id);
+        if(userRepository.findById(id).isEmpty()){
+            throw new UserNotFoundException("User not found");
+        } else {
+            userRepository.deleteById(id);
+        }
     }
 }
